+++ SourceryTests/Parsing/FileParserSpec.swift
@@ -484,6 +484,23 @@ class FileParserSpec: QuickSpec {
                                           ]))
                     }
 
+                    it("extracts default values for asssociated values") {
+                        expect(parse("enum Foo { case optionA(Int = 1, named: Float = 42.0, _: Bool = false); case optionB(Bool = true) }"))
+                        .to(equal([
+                            Enum(name: "Foo", accessLevel: .internal, isExtension: false, inheritedTypes: [], cases:
+                                [
+                                    EnumCase(name: "optionA", associatedValues: [
+                                        AssociatedValue(localName: nil, externalName: "0", typeName: TypeName("Int"), defaultValue: "1"),
+                                        AssociatedValue(localName: "named", externalName: "named", typeName: TypeName("Float"), defaultValue: "42.0"),
+                                        AssociatedValue(localName: nil, externalName: "2", typeName: TypeName("Bool"), defaultValue: "false")
+                                        ]),
+                                    EnumCase(name: "optionB", associatedValues: [
+                                    AssociatedValue(localName: nil, externalName: nil, typeName: TypeName("Bool"), defaultValue: "true")
+                                    ])
+                                ])
+                        ]))
+                    }
+
                     context("given associated value with its type existing") {
 
                         it("extracts associated value's type") {