+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 - Added `optional` filter for variables
 - Added `json` filter to output raw JSON objects 
+- Added `.defaultValue` to `AssociatedValue`
 
 ### Internal Changes
 
@@ -16,6 +17,7 @@
 
 - Fixed expansion of undefined environment variables (now consistent with command line behaviour, where such args are empty strings)
 - Fixed a bug in inferring extensions of Dictionary and Array types
+- Fixed a bug that was including default values as part of AssociatedValues type names
 
 ## 0.17.0